pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fretish-robot"
version = "0.0.1"
authors = [
    { name = "Stefan Kraus", email = "stefan.kraus@ul.com" },
]

description = "A package to support FRET requirement usage in Robot Framework"
requires-python = ">=3.9"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]

dependencies = [
    "robotframework==7.0",
    "jinja2==3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
fret-to-robot = "fretish_robot.cli:main"

[project.urls]
repository = "https://github.com/ZEISS/fretish_robot"

[tool.hatch.build.targets.wheel]
packages = ["src/fretish_robot"]
